All manager API impls use the same jax-rs application class now
