+++ manager/api/rest-impl/src/main/java/io/apiman/manager/api/rest/impl/ApiManagerApplication.java
@@ -16,7 +16,13 @@
 
 package io.apiman.manager.api.rest.impl;
 
+import io.apiman.manager.api.config.Version;
 import io.apiman.manager.api.exportimport.manager.ExportImportManager;
+import io.apiman.manager.api.rest.exceptions.mappers.RestExceptionMapper;
+import io.swagger.jaxrs.config.BeanConfig;
+
+import java.util.HashSet;
+import java.util.Set;
 
 import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
@@ -37,10 +43,36 @@ public class ApiManagerApplication extends Application {
     @Inject
     ExportImportManager manager;
 
+    private Set<Class<?>> classes = new HashSet<>();
+
     /**
      * Constructor.
      */
     public ApiManagerApplication() {
+        //add swagger 2.0 config
+        BeanConfig beanConfig = new BeanConfig();
+        beanConfig.setVersion(new Version().getVersionString());
+        beanConfig.setBasePath("/apiman"); //$NON-NLS-1$
+        beanConfig.setResourcePackage("io.apiman.manager.api.rest"); //$NON-NLS-1$
+        beanConfig.setDescription("The API Manager REST API is used by the API Manager UI to get stuff done. You can use it to automate any API Management task you wish. For example, create new Organizations, Plans, Clients, and APIs.");
+        beanConfig.setScan(true);
+
+        classes.add(SystemResourceImpl.class);
+        classes.add(SearchResourceImpl.class);
+        classes.add(RoleResourceImpl.class);
+        classes.add(UserResourceImpl.class);
+        classes.add(OrganizationResourceImpl.class);
+        classes.add(PolicyDefinitionResourceImpl.class);
+        classes.add(GatewayResourceImpl.class);
+        classes.add(PluginResourceImpl.class);
+        classes.add(ActionResourceImpl.class);
+        classes.add(DownloadResourceImpl.class);
+
+        //add swagger 2.0 resource
+        classes.add(io.swagger.jaxrs.listing.ApiListingResource.class);
+        classes.add(io.swagger.jaxrs.listing.SwaggerSerializers.class);
+
+        classes.add(RestExceptionMapper.class);
     }
     
     @PostConstruct
@@ -49,5 +81,9 @@ public class ApiManagerApplication extends Application {
             manager.doImportExport();
         }
     }
-    
+
+    @Override
+    public Set<Class<?>> getClasses() {
+        return classes;
+    }
 }
